Update to support fallback behaviour for Reveal detection. If Reveal 2 is installed, we'll use that otherwise fall back to 1.x.
